Add child process handler
Adds child process handlers which starts and terminates the process which is used
to automatically query the web server for new episodes. This should remove the
dependence on the cron system that would be normally OS bound. However spawning a
child process WILL start another V8 instance and does come with some weight.